Simplify copying and avoid excessive permission damage
chmod -R changes permissions for all files and subdirectories to
rwxrwxrwx which certainly is excessive. It's even doubtful, that
directories need to be writable for all, the only write access seen
to them so far was by root. For now keep directories writable for all
but keep files with their original permissions.

Signed-off-by: Guennadi Liakhovetski <g.liakhovetski@gmx.de>